ai_data_visualisation_agent/requirements.txt
openai>=1.0.0
pandas>=2.0.0
plotly>=5.0.0
python-dotenv>=1.0.0
